Remove duplicated platform field from oci index
It is desirable to remove Platform from distribution.Descriptor because it doesn't really belong there. However this would be a further breaking change because the References() call would no longer be returning plaform information when it reurns descriptors of manifests, which is started to for OCI Indices after 99657b7 and this feature was added to Docker Manifest Lists in 6a37d13. I don't want to take away something people clearly want.

Signed-off-by: Bracken Dawson <abdawson@gmail.com>