reference: fix broken alias for DomainRegexp
An incorrect alias snuck into b4659b171658ce1b3837d108b8c4e6f6b1bf72a0,
and DomainRegexp was aliased to the regex for digests (DigestRegexp).

This didn't affect this repository, as it didn't use the aliases and migrated
to the new module, but does affect user of the old module that depend on the
aliases.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>